Bug fix
Fixed Bug detail: If there is an ephemeral volume  in AMI's Block device mapping, it was throwing an error. Information of block device mapping is taken for deleting snapshots associated with the AMI. Ephermal volume mapping is skipped as there will be no snapshot associated with it. 

Thanks Vinod Purushothaman (purushothaman.vinod@gmail.com) for reporting the bug.